+++ Runtime/GeminiLiveForUnity/WezzelNL/Util/AsyncDispatcher.cs
@@ -3,6 +3,9 @@ using System.Runtime.ExceptionServices;
 using System.Threading;
 using System.Threading.Tasks;
 using UnityEngine;
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
 
 namespace WezzelNL.Util
 {
@@ -17,27 +20,11 @@ namespace WezzelNL.Util
             {
                 await Awaitable.MainThreadAsync();
                 var exception = ExceptionDispatchInfo.Capture(e);
-                new GameObject("Temp").AddComponent<AsyncExceptionThrower>().Throw(exception);
-            }
-        }
-    }
-
-    public class AsyncExceptionThrower : MonoBehaviour
-    {
-        private ExceptionDispatchInfo exception;
-        private bool canThrow = false;
-        public void Throw(ExceptionDispatchInfo e)
-        {
-            exception = e;
-            canThrow = true;
-        }
-
-        private void Update()
-        {
-            if (canThrow)
-            {
-                DestroyImmediate(gameObject);
-                exception.Throw();
+                #if UNITY_EDITOR
+                EditorApplication.delayCall += exception.Throw;
+                #else
+                Debug.LogException(e);
+                #endif
             }
         }
     }